Moved middleware logic to controllers from routes
