+++ backend/routes/posts.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const multer = require('multer');
+const PostsController = require('../controllers/posts');
 
-const Post = require('../models/post.js');
 const checkAuth = require('../middleware/check-auth');
 
 const router = express.Router();
@@ -26,123 +26,14 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post('', checkAuth, multer({storage: storage}).single('image'), (req, res, next) => {
-  const url = req.protocol + '://' + req.get('host');
-  const post = new Post({
-    title: req.body.title,
-    content: req.body.content,
-    imagePath:  url + '/images/' + req.file.filename,
-    creator: req.userData.userId
-  });
-  post.save().then(createdPost => {
-    res.status(201).json({
-      message: 'All done',
-      post: {
-        //...createdPost, next gen stuff, copies first then replaces following
-        id: createdPost._id,
-        title: createdPost.title,
-        content: createdPost.content,
-        imagePath: createdPost.imagePath
-      }
-    });
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Creating post failed!"
-    })
-  });
-  console.log(post);
-});
+router.post('', checkAuth, multer({storage: storage}).single('image'), PostsController.createPost);
 
-router.put('/:id', checkAuth, multer({storage: storage}).single('image'), (req,res,next) => {
-  let imagePath = req.body.imagePath;
-  if(req.file) {
-    const url = req.protocol + '://' + req.get('host');
-    imagePath = url + '/images/' + req.file.filename;
-  }
-  const post = new Post({
-    _id: req.body.id,
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: imagePath,
-    creator: req.userData.userId // note that that req has null value on this field cause of security.
-  });
+router.put('/:id', checkAuth, multer({storage: storage}).single('image'), PostsController.updatePost);
 
-  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then( result => {
-    if (result.nModified > 0) {
-      console.log('Sucess update on: ' + post);
-      res.status(200).json({ message: "Update successful!" });
-    } else {
-      console.log('Not authorized update on: ' + post);
-      res.status(401).json({ message: 'Not authorized!' })
-    }
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Couldn't update post!"
-    })
-  })
-});
+router.get('', PostsController.getPosts);
 
-router.get('', (req, res, next) => {
-  const pageSize = +req.query.pagesize;
-  const currentPage = +req.query.page;
-  const postQuery = Post.find();
-  let fetchedPosts;
+router.get('/:id', PostsController.getPost);
 
-  if(pageSize && currentPage) {
-    postQuery
-    .skip(pageSize * (currentPage - 1))
-    .limit(pageSize);
-  }
-
-  postQuery.then(documents => {
-    fetchedPosts = documents;
-    return Post.count()
-  }).then(count => {
-    res.status(200).json({
-      message : 'Fetched posts sucess!',
-      posts: fetchedPosts,
-      maxPosts: count
-    });
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Fetching posts failed!"
-    })
-  });
-});
-
-router.get('/:id', (req, res, next) => {
-  Post.findById(req.params.id).then(post => {
-    if(post) {
-      res.status(200).json(post);
-    } else {
-      res.status(404).json({message: 'Post not found!'});
-    }
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Fetching posts failed!"
-    })
-  });
-});
-
-router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
-    if (result.n > 0) {
-      console.log('Sucessful delete on: ');
-      res.status(200).json({message: 'Post Deleted!'});
-    } else {
-      console.log('Not authorized delete!');
-      res.status(401).json({ message: 'Not authorized!' })
-    }
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Fetching posts failed!"
-    })
-  });
-});
+router.delete('/:id', checkAuth, PostsController.deletePost);
 
 module.exports = router;
